backend/requirements.txt
flask
flask_sqlalchemy
werkzeug
flask-cors
pymysql
